refactor: restructure model mixins to improve code organization
- Split SoftDeleteMixin into separate file for better logical organization
- Remove duplicated mixins.py file causing 87.5% code duplication
- Update imports to use new file structure
- Add backward compatibility via re-exports in base_models
- Improve package docstrings for better code navigation
