+++ backend/models/base.py
@@ -1,10 +1,10 @@
 """Base model module.
 
-This module re-exports base model classes from base_models package.
+This module re-exports base model classes from core package.
 """
 
-from backend.models.base_models.base import Base, HasId
-from backend.models.base_models.mixins import SoftDeleteMixin, TimestampMixin
+from backend.models.core.base import Base, HasId
+from backend.models.mixins import SoftDeleteMixin, TimestampMixin
 
 __all__ = [
     'Base',
