+++ backend/models/base_models/__init__.py
@@ -1,5 +1,10 @@
-"""Base models package.
+"""Base models package for backward compatibility.
 
-This package contains base classes and mixins for models.
-It's separated to avoid circular imports.
+This package re-exports functionality from the new model structure.
+It's maintained for backward compatibility during refactoring period.
 """
+
+from backend.models.core.base import Base, HasId
+from backend.models.mixins import SoftDeleteMixin, TimestampMixin
+
+__all__ = ['Base', 'HasId', 'SoftDeleteMixin', 'TimestampMixin']
