+++ backend/models/mixins/__init__.py
@@ -4,6 +4,7 @@ This package contains mixins for models.
 It's separated to avoid circular imports.
 """
 
-from backend.models.mixins.timestamps import SoftDeleteMixin, TimestampMixin
+from backend.models.mixins.soft_delete import SoftDeleteMixin
+from backend.models.mixins.timestamps import TimestampMixin
 
 __all__ = ['TimestampMixin', 'SoftDeleteMixin']
